requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# HTTP requests
requests

# Web scraping
selenium

# Database - sqlite3 is built into Python
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.27.0
aiohttp>=3.9.0